+++ cnxarchive/views/exports.py
@@ -28,7 +28,8 @@ from .. import database
 from ..database import SQL, get_tree, get_collated_content
 
 from ..utils import (
-    slugify, fromtimestamp, join_ident_hash, split_ident_hash
+    slugify, fromtimestamp, join_ident_hash, split_ident_hash,
+    portaltype_to_mimetype
     )
 from .content import get_content_metadata
 
@@ -45,6 +46,30 @@ class ExportError(Exception):
     """Used as catchall for other export errors."""
 
     pass
+
+
+def get_export_allowable_types(cursor, exports_dirs, id, version):
+    """Return export types."""
+    request = get_current_request()
+
+    for type_name, type_info in request.registry.settings['_type_info']:
+        try:
+            (filename, mimetype, file_size, file_created, state, file_content
+             ) = get_export_file(cursor, id, version, type_name, exports_dirs)
+            yield {
+                'format': type_info['user_friendly_name'],
+                'filename': filename,
+                'size': file_size,
+                'created': file_created and file_created.isoformat() or None,
+                'state': state,
+                'details': type_info['description'],
+                'path': request.route_path(
+                    'export', ident_hash=join_ident_hash(id, version),
+                    type=type_name, ignore=u'/{}'.format(filename))
+                }
+        except ExportError as e:  # noqa
+            # Some other problem, skip it
+            pass
 
 
 # ######### #
@@ -79,30 +104,6 @@ def get_export(request):
     resp.content_disposition = u'attached; filename={}'.format(filename)
     resp.body = file_content
     return resp
-
-
-def get_export_allowable_types(cursor, exports_dirs, id, version):
-    """Return export types."""
-    request = get_current_request()
-
-    for type_name, type_info in request.registry.settings['_type_info']:
-        try:
-            (filename, mimetype, file_size, file_created, state, file_content
-             ) = get_export_file(cursor, id, version, type_name, exports_dirs)
-            yield {
-                'format': type_info['user_friendly_name'],
-                'filename': filename,
-                'size': file_size,
-                'created': file_created and file_created.isoformat() or None,
-                'state': state,
-                'details': type_info['description'],
-                'path': request.route_path(
-                    'export', ident_hash=join_ident_hash(id, version),
-                    type=type_name, ignore=u'/{}'.format(filename))
-                }
-        except ExportError as e:  # noqa
-            # Some other problem, skip it
-            pass
 
 
 def get_export_file(cursor, id, version, type, exports_dirs):
