+++ cnxarchive/views/views_helpers.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+# ###
+# Copyright (c) 2013, Rice University
+# This software is subject to the provisions of the GNU Affero General
+# Public License version 3 (AGPLv3).
+# See LICENCE.txt for details.
+# ###
+"""Helpers Used in Multiple Views."""
+import os
+import json
+import logging
+from datetime import datetime, timedelta
+from re import compile
+
+import psycopg2
+import psycopg2.extras
+from cnxepub.models import flatten_tree_to_ident_hashes
+from lxml import etree
+from pytz import timezone
+from pyramid import httpexceptions
+from pyramid.settings import asbool
+from pyramid.threadlocal import get_current_registry, get_current_request
+from pyramid.view import view_config
+
+from .. import config
+from .. import cache
+from ..database import SQL, get_module_can_publish
+from ..utils import fromtimestamp, split_ident_hash, portaltype_to_mimetype
+
+logger = logging.getLogger('cnxarchive')
+
+
+# #################### #
+#   Helper functions   #
+# #################### #
+
+
+def get_uuid(shortid):
+    settings = get_current_registry().settings
+    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
+        with db_connection.cursor() as cursor:
+            cursor.execute(SQL['get-module-uuid'], {'id': shortid})
+            try:
+                return cursor.fetchone()[0]
+            except (TypeError, IndexError,):  # None returned
+                logger.debug("Short ID was supplied and could not discover "
+                             "UUID.")
+                raise httpexceptions.HTTPNotFound()
+
+
+def get_latest_version(uuid_):
+    settings = get_current_registry().settings
+    with psycopg2.connect(settings[config.CONNECTION_STRING]) as db_connection:
+        with db_connection.cursor() as cursor:
+            cursor.execute(SQL['get-module-versions'], {'id': uuid_})
+            try:
+                return cursor.fetchone()[0]
+            except (TypeError, IndexError,):  # None returned
+                raise httpexceptions.HTTPNotFound()
